Animation
Created animation for character, by designing a sprite sheet for walking, idle and climbing. For each action I gave each function a bool whether it is true or false for each action to happen and when. Also gave player an action and added player layer so that the character can appear infront of the background and lands neatly on the foreground.