Colliders
I added a rigidbody 2D and a capsule collider 2D, to make it look more smoother rather than using the box or circle collider. As for the foreground, the make interection with the player, since we are adding gravity to the player, we added a tilemap collider for the foreground as the background doesn't need the same function as there is no interection with the player for now. For the tilemap collider I added as well a composite collider to fill in the micro gaps in between tiles to make the gameplay smoother and player hass less chances to get stuck in between these cracks.  I also added a rigidbody 2D for the foreground and made it static, as it doesn't need to move around in the game unlike the player.